update GPU package support for r561 from svn.

+++ lib/gpu/atomic_gpu_memory.cpp
@@ -116,7 +116,7 @@ int * AtomicGPUMemoryT::reset_nbors(const int nall, const int inum, int *ilist,
 
   nbor->get_host(inum,ilist,numj,firstneigh,block_size());
 
-  int bytes=atom->gpu_bytes()+nbor->gpu_bytes();
+  double bytes=atom->gpu_bytes()+nbor->gpu_bytes();
   if (bytes>_max_an_bytes)
     _max_an_bytes=bytes;
   
@@ -147,7 +147,7 @@ inline void AtomicGPUMemoryT::build_nbor_list(const int inum,
   nbor->build_nbor_list(inum, host_inum, nall, *atom, boxlo, boxhi, tag,
                         nspecial, special, success, mn);
 
-  int bytes=atom->gpu_bytes()+nbor->gpu_bytes();
+  double bytes=atom->gpu_bytes()+nbor->gpu_bytes();
   if (bytes>_max_an_bytes)
     _max_an_bytes=bytes;
 }
